requirements.txt
streamlit==1.28.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.0
typing-extensions==4.8.0
requests==2.31.0
validators==0.22.0
json5==0.9.14